pyproject.toml
[project]
name = "autoagent"
version = "0.1.0"
description = ""
authors = [
    {name = "Dinesh Ram R",email = "dineshram0212@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai (>=1.75.0,<2.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "faiss-cpu (>=1.10.0,<2.0.0)",
    "sentence-transformers (>=4.1.0,<5.0.0)",
    "numpy (>=2.2.5,<3.0.0)",
    "chromadb (>=1.0.7,<2.0.0)"
]

[tool.poetry]
packages = [{include = "autoagent", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
